Improved the wasteagram post form upload button and custom TextStyle class.
While the previous icon button was able to upload correctly, the tappable area was only a fraction of the rectangle that visually represented the upload button.  I was able to create a FloatingActionButton with the correct tappable area and an icon that can better scale with different sized screens.

Added a custom TextStyle class.  While not as helpful here, it seems to be a common convention for flutter users.