Added Firebase Analytics to the project.  Fixed the button icon for my list view page.
The analytics will track when the app is opened and what posts are opened.  The observer attached to the navigator will track screens as well.  This is very crude.  Events containing the exact details of every post selected could be created, but this helps me see how information is collected and stored in Firebase.